Finish Macro Implementation
Bug fixes as well as checking of macro features. The  function ended up not being a native Windows function, which seems a bit illogical given the malware description. The implementation I found online also doesn't work as intended with the files I've been able to create, so further work is necessary